Added detection of electron in boot.
Added $tw.electron variable, which is non-null if TW is being run on electron.

Under electron, the main process runs the node.js server, and the browser process runs the browser tiddlywiki.

For this to work, $tw.node is forcedly set to null if $tw.electron and $tw.browser are both non-null.